refactor(checker): split lib_resolution integration tests into sibling module
The merge with main pushed lib_resolution.rs to 2007 lines (ceiling
2000; main was already at 1998). Move the inline integration_tests
module (430 lines, crate-path-only imports) to
lib_resolution_integration_tests.rs instead of shaving comments —
restores ~430 lines of headroom for both sides.
